Crate icu_collections

Efficient collections for Unicode data.

This module is published as its own crate (icu_collections) and as part of the icu crate. See the latter for more details on the ICU4X project.

ICU4X CodePointTrie provides a read-only view of CodePointTrie data that is exported from ICU4C. Detailed information about the design of the data structure can be found in the documentation for the CodePointTrie struct.

ICU4X CodePointInversionList provides necessary functionality for highly efficient querying of sets of Unicode characters. It is an implementation of the existing ICU4C UnicodeSet API.

ICU4X Char16Trie provides a data structure for a space-efficient and time-efficient lookup of sequences of 16-bit units (commonly but not necessarily UTF-16 code units) which map to integer values. It is an implementation of the existing ICU4C UCharsTrie / ICU4J CharsTrie API.

Modules

char16trie

This module provides a data structure for a space-efficient and time-efficient lookup of sequences of 16-bit units (commonly but not necessarily UTF-16 code units) which map to integer values.

codepointinvlist

This module provides necessary functionality for highly efficient querying of sets of Unicode characters.

codepointtrie

This module provides a data structure for an time-efficient lookup of values associated to code points.