icu_collections::codepointinvlist

Module cpinvlist

Structs

CodePointInversionList

A membership wrapper for CodePointInversionList.

CodePointInversionListULE

VarULE type for CodePointInversionList. See CodePointInversionList for documentation.

Constants

ALL_VEC 🔒

Represents the inversion list for all of the code points in the Unicode range.

BMP_INV_LIST_VEC 🔒

Represents the inversion list for a set of all code points in the Basic Multilingual Plane.

BMP_MAX 🔒

Represents the end code point of the Basic Multilingual Plane range, starting from code point 0, inclusive

__IMPL_VarULE_FOR_CodePointInversionListULE_ULE_SIZE 🔒