icu_collections::codepointtrie

Module cptrie

Macros

impl_primitive_trie_value 🔒

w 🔒

Structs

CodePointMapRange

Represents a range of consecutive code points sharing the same value in a code point map.

CodePointMapRangeIterator

A custom Iterator type specifically for a code point trie that returns CodePointMapRanges.

CodePointTrie

This struct represents a de-serialized CodePointTrie that was exported from ICU binary data.

CodePointTrieHeader

This struct contains the fixed-length header fields of a CodePointTrie.

FastCodePointTrie

Type-safe wrapper for a fast trie guaranteeing the the getters don’t branch on the trie type and for guarenteeing that get16 is branchless in release builds.

SmallCodePointTrie

Type-safe wrapper for a small trie guaranteeing the the getters don’t branch on the trie type.

TypedCodePointTrieError

Error indicating that the TrieType of an untyped trie does not match the requested typed trie type.

Enums

TrieType

The type of trie represents whether the trie has an optimization that would make it smaller or faster.

Typed

Holder for either fast or small trie with the trie type encoded into the Rust type.

Traits

AbstractCodePointTrie

Trait for writing trait bounds for monomorphizing over either CodePointTrie, FastCodePointTrie, or SmallCodePointTrie.

Seal 🔒

For sealing TypedCodePointTrie

TrieValue

A trait representing the values stored in the data array of a CodePointTrie. This trait is used as a type parameter in constructing a CodePointTrie.

TypedCodePointTrie

Trait for writing trait bounds for monomorphizing over either FastCodePointTrie or SmallCodePointTrie.

Functions

maybe_filter_value 🔒

Helper function used by [get_range]. Converts occurrences of trie’s null value into the provided null_value.