icu_collections::codepointtrie::cptrie

Function maybe_filter_value

fn maybe_filter_value<T: TrieValue>(
    value: T,
    trie_null_value: T,
    null_value: T,
) -> T

Helper function used by [get_range]. Converts occurrences of trie’s null value into the provided null_value.

Note: the ICU version of this helper function uses a ValueFilter function to apply a transform on a non-null value. But currently, this implementation stops short of that functionality, and instead leaves the non-null trie value untouched. This is equivalent to having a ValueFilter function that is the identity function.