icu_collections::codepointtrie::cptrie

Struct FastCodePointTrie

#[repr(transparent)]
pub struct FastCodePointTrie<'trie, T: TrieValue> {
    inner: CodePointTrie<'trie, T>,
}

Type-safe wrapper for a fast trie guaranteeing the the getters don’t branch on the trie type and for guarenteeing that get16 is branchless in release builds.

Fields
inner: CodePointTrie<'trie, T>