icu_collections::codepointtrie

Module iter

Structs

CharIndicesWithTrie

Iterator over str by char and TrieValue.

CharIndicesWithTrieDefaultForAscii

Iterator over str by char and TrieValue.

CharIterWithTrie

Wraps an Iterator<Item = char> with a reference to an AbstractCodePointTrie.

CharsWithTrie

Iterator over str by char and TrieValue.

CharsWithTrieDefaultForAscii

Iterator over str by char and TrieValue but the trie value for ASCII is V::default() instead of reading from the trie. (V::default() can be optimized on at compile time while reading the trie’s default value is a run-time operation.)

Latin1CharIndicesWithTrie

Iterator over str by char and TrieValue.

Latin1CharsWithTrie

Iterator over Latin1 [u8] by char and TrieValue.

Traits

CharsWithTrieDefaultForAsciiEx

Adds convenience methods to str.

CharsWithTrieEx

Adds convenience methods to str.

Latin1CharsWithTrieEx

Adds convenience methods to [u8].

WithTrie

Provides a trie accessor for types (likely iterators) that are holding a reference to a type that implements AbstractCodePointTrie.