icu_collections::codepointtrie

Module planes

Sample data for CodePointTrie that returns the code point’s plane number.

Functions

get_planes_trie

Return a CodePointTrie that returns the Unicode plane number, an integer from 0-16 inclusive, for each code point. This CodePointTrie does not actually represent any Unicode property, but it is provided in case it is useful to users of CodePointTrie for testing or other purposes. See https://www.unicode.org/glossary/#plane.