Function icu_collections::codepointtrie::planes::get_planes_trie

pub fn get_planes_trie() -> CodePointTrie<'static, u8>

Return a CodePointTrie that returns the Unicode plane number, an integer from 0-16 inclusive, for each code point. This CodePointTrie does not actually represent any Unicode property, but it is provided in case it is useful to users of CodePointTrie for testing or other purposes. See https://www.unicode.org/glossary/#plane.