icu_collections
Module char16trie

This module provides a data structure for a space-efficient and time-efficient lookup of sequences of 16-bit units (commonly but not necessarily UTF-16 code units) which map to integer values.

It is an implementation of the existing ICU4C UCharsTrie / ICU4J CharsTrie API.

Architecture

ICU4X Char16Trie is designed to provide a read-only view of UCharsTrie data that is exported from ICU4C.

Examples

Querying a Char16Trie

use icu::collections::char16trie::{Char16Trie, TrieResult};
use zerovec::ZeroVec;

// A Char16Trie containing the ASCII characters mapping 'a' to 1 and 'ab'
// to 100.
let trie_data = [48, 97, 176, 98, 32868];
let trie = Char16Trie::new(ZeroVec::from_slice_or_alloc(&trie_data));

let mut iter = trie.iter();
let res = iter.next('a');
assert_eq!(res, TrieResult::Intermediate(1));
let res = iter.next('b');
assert_eq!(res, TrieResult::FinalValue(100));
let res = iter.next('c');
assert_eq!(res, TrieResult::NoMatch);

Structs

Char16Trie

This struct represents a de-serialized Char16Trie that was exported from ICU binary data.

Char16TrieIterator

This struct represents an iterator over a Char16Trie.

Enums

TrieResult

An enum representing the return value from a lookup in Char16Trie.