icu_collections::char16trie

Module trie

Macros

trie_unwrap 🔒

A macro that takes an Option argument and either unwraps it if it has a value or causes the function to return TrieResult::NoMatch if there is no value. This could perhaps be done with std::ops::Try once stabilized.

Structs

Char16Trie

This struct represents a de-serialized Char16Trie that was exported from ICU binary data.

Char16TrieIterator

This struct represents an iterator over a Char16Trie.

Enums

TrieResult

An enum representing the return value from a lookup in Char16Trie.

Constants

MAX_BRANCH_LINEAR_SUB_NODE_LENGTH 🔒

MAX_LINEAR_MATCH_LENGTH 🔒

MAX_ONE_UNIT_DELTA 🔒

MAX_ONE_UNIT_NODE_VALUE 🔒

MAX_ONE_UNIT_VALUE 🔒

MIN_LINEAR_MATCH 🔒

MIN_TWO_UNIT_DELTA_LEAD 🔒

MIN_TWO_UNIT_NODE_VALUE_LEAD 🔒

MIN_TWO_UNIT_VALUE_LEAD 🔒

MIN_VALUE_LEAD 🔒

NODE_TYPE_MASK 🔒

THREE_UNIT_DELTA_LEAD 🔒

THREE_UNIT_NODE_VALUE_LEAD 🔒

THREE_UNIT_VALUE_LEAD 🔒

VALUE_IS_FINAL 🔒

Functions

skip_node_value 🔒

skip_value 🔒

u16_lead 🔒

u16_tail 🔒