icu_collections

Module codepointinvliststringlist

This module provides functionality for querying of sets of Unicode code points and strings.

It depends on CodePointInversionList to efficiently represent Unicode code points, while it also maintains a list of strings in the set.

It is an implementation of the existing ICU4C UnicodeSet API.

Re-exports

pub use CodePointInversionListAndStringListError as Error;

Structs

CodePointInversionListAndStringList

A data structure providing a concrete implementation of a UnicodeSet (which represents a set of code points and strings) using an inversion list for the code points and a simple list-like structure to store and iterate over the strings.

CodePointInversionListAndStringListULE

VarULE type for CodePointInversionListAndStringList. See CodePointInversionListAndStringList for documentation.

Enums

CodePointInversionListAndStringListError

Custom Errors for CodePointInversionListAndStringList.