icu_collections::codepointinvliststringlist

Struct CodePointInversionListAndStringList

pub struct CodePointInversionListAndStringList<'data> { /* private fields */ }

A data structure providing a concrete implementation of a UnicodeSet (which represents a set of code points and strings) using an inversion list for the code points and a simple list-like structure to store and iterate over the strings.

Implementations

impl<'data> CodePointInversionListAndStringList<'data>

pub fn try_from( cp_inv_list: CodePointInversionList<'data>, str_list: VarZeroVec<'data, str>, ) -> Result<Self, CodePointInversionListAndStringListError>

Returns a new CodePointInversionListAndStringList from both a CodePointInversionList for the code points and a VarZeroVec<str> of strings.

pub fn size(&self) -> usize

Returns the number of elements in this set (its cardinality). Note than the elements of a set may include both individual codepoints and strings.

pub fn has_strings(&self) -> bool

Return true if this set contains multi-code point strings or the empty string.

pub fn contains(&self, s: &str) -> bool

Examples

use icu::collections::codepointinvlist::CodePointInversionList;
use icu::collections::codepointinvliststringlist::CodePointInversionListAndStringList;
use zerovec::VarZeroVec;

let cp_slice = &[0, 0x1_0000, 0x10_FFFF, 0x11_0000];
let cp_list =
   CodePointInversionList::try_clone_from_inversion_list_slice(cp_slice).unwrap();
let str_slice = &["", "bmp_max", "unicode_max", "zero"];
let str_list = VarZeroVec::<str>::from(str_slice);

let cpilsl = CodePointInversionListAndStringList::try_from(cp_list, str_list).unwrap();

assert!(cpilsl.contains("bmp_max"));
assert!(cpilsl.contains(""));
assert!(cpilsl.contains("A"));
assert!(cpilsl.contains("แ"));  // U+1254 ETHIOPIC SYLLABLE QHEE
assert!(!cpilsl.contains("bazinga!"));

pub fn contains32(&self, cp: u32) -> bool

Examples

use icu::collections::codepointinvlist::CodePointInversionList;
use icu::collections::codepointinvliststringlist::CodePointInversionListAndStringList;
use zerovec::VarZeroVec;

let cp_slice = &[0, 0x80, 0xFFFF, 0x1_0000, 0x10_FFFF, 0x11_0000];
let cp_list =
    CodePointInversionList::try_clone_from_inversion_list_slice(cp_slice).unwrap();
let str_slice = &["", "ascii_max", "bmp_max", "unicode_max", "zero"];
let str_list = VarZeroVec::<str>::from(str_slice);

let cpilsl = CodePointInversionListAndStringList::try_from(cp_list, str_list).unwrap();

assert!(cpilsl.contains32(0));
assert!(cpilsl.contains32(0x0042));
assert!(!cpilsl.contains32(0x0080));

pub fn contains_char(&self, ch: char) -> bool

Examples

use icu::collections::codepointinvlist::CodePointInversionList;
use icu::collections::codepointinvliststringlist::CodePointInversionListAndStringList;
use zerovec::VarZeroVec;

let cp_slice = &[0, 0x1_0000, 0x10_FFFF, 0x11_0000];
let cp_list =
   CodePointInversionList::try_clone_from_inversion_list_slice(cp_slice).unwrap();
let str_slice = &["", "bmp_max", "unicode_max", "zero"];
let str_list = VarZeroVec::<str>::from(str_slice);

let cpilsl = CodePointInversionListAndStringList::try_from(cp_list, str_list).unwrap();

assert!(cpilsl.contains_char('A'));
assert!(cpilsl.contains_char('แ'));  // U+1254 ETHIOPIC SYLLABLE QHEE
assert!(!cpilsl.contains_char('\u{1_0000}'));
assert!(!cpilsl.contains_char('๐จซ'));  // U+1FA2B NEUTRAL CHESS TURNED QUEEN

pub fn code_points(&self) -> &CodePointInversionList<'data>

Access the underlying CodePointInversionList.

pub fn strings(&self) -> &VarZeroSlice<str>

Access the contained strings.