Module icu_collections::codepointtrie::cptrie

Macros

impl_primitive_trie_value 🔒

Structs

CodePointMapRange

Represents a range of consecutive code points sharing the same value in a code point map. The start and end of the interval is represented as a RangeInclusive<u32>, and the value is represented as T.

CodePointMapRangeIterator

A custom Iterator type specifically for a code point trie that returns CodePointMapRanges.

CodePointTrie

This struct represents a de-serialized CodePointTrie that was exported from ICU binary data.

CodePointTrieHeader

This struct contains the fixed-length header fields of a CodePointTrie.

Enums

TrieType

The type of trie represents whether the trie has an optimization that would make it smaller or faster.

Traits

TrieValue

A trait representing the values stored in the data array of a CodePointTrie. This trait is used as a type parameter in constructing a CodePointTrie.

Functions

maybe_filter_value 🔒

Helper function used by [get_range]. Converts occurrences of trie’s null value into the provided null_value.