Module icu_collections::codepointtrie::impl_const

Constants

BMP_INDEX_LENGTH

The length of the BMP index table. 1024=0x400

CODE_POINT_MAX

CP_PER_INDEX_2_ENTRY

Number of code points per index-2 table entry. 512=0x200

ERROR_VALUE_NEG_DATA_OFFSET

Offset from dataLength (to be subtracted) for fetching the value returned for out-of-range code points and ill-formed UTF-8/16.

FAST_TYPE_DATA_BLOCK_LENGTH

Number of entries in a data block for code points below the fast limit. 64=0x40

FAST_TYPE_DATA_MASK

Mask for getting the lower bits for the in-fast-data-block offset.

FAST_TYPE_FAST_INDEXING_MAX

Fast indexing limit for “fast”-type trie

FAST_TYPE_SHIFT

HIGH_VALUE_NEG_DATA_OFFSET

Offset from dataLength (to be subtracted) for fetching the value returned for code points highStart..U+10FFFF.

INDEX_2_BLOCK_LENGTH

Number of entries in an index-2 block. 32=0x20

INDEX_2_MASK

Mask for getting the lower bits for the in-index-2-block offset.

INDEX_3_BLOCK_LENGTH

Number of entries in an index-3 block. 32=0x20

INDEX_3_MASK

Mask for getting the lower bits for the in-index-3-block offset.

OMITTED_BMP_INDEX_1_LENGTH

Number of index-1 entries for the BMP. (4) This part of the index-1 table is omitted from the serialized form.

SHIFT_1

Shift size for getting the index-1 table offset.

SHIFT_1_2

Difference between two shift sizes, for getting an index-1 offset from an index-2 offset. 5=14-9

SHIFT_2

Shift size for getting the index-2 table offset.

SHIFT_2_3

Difference between two shift sizes, for getting an index-2 offset from an index-3 offset. 5=9-4

SHIFT_3

Shift size for getting the index-3 table offset.

SMALL_DATA_BLOCK_LENGTH

Number of entries in a small data block. 16=0x10

SMALL_DATA_MASK

Mask for getting the lower bits for the in-small-data-block offset.

SMALL_INDEX_LENGTH

SMALL_LIMIT

SMALL_TYPE_FAST_INDEXING_MAX

Fast indexing limit for “small”-type trie